backend/requirements.txt
fastapi==0.128.5
sqlmodel==0.0.32
uvicorn==0.40.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
python-multipart==0.0.22
pydantic-settings==2.12.0
openai==1.57.4